requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-core==1.35.0
azure-identity==1.24.0
bcrypt==3.2.2
boto3==1.40.33
# boto3==1.40.6
botocore==1.40.33
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.1.5
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
msal==1.33.0
msal-extensions==1.3.1
mssql-python==0.9.0
numpy==2.3.3
pandas==2.3.2
passlib==1.7.4
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.4
pyroaring==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
s3transfer==0.14.0
# s3transfer==0.13.9
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.48.0
strictyaml==1.7.3
tenacity==9.1.2
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
